LPS-55043 Move
